lesson7 refactor: add ChatsViewContainer
