+++ src/chat/ChatsView.js
@@ -1,21 +1,8 @@
 import MessageList from "../MessageList.js";
 import MessageForm from "../MessageForm.js";
-import { useCallback, useEffect, useMemo } from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import { Grid, Button } from "@material-ui/core";
 import ChatList from "../chatList/ChatList";
-import { useHistory, useParams } from "react-router";
-import { useDispatch, useSelector } from "react-redux";
-import {
-  chatListAddChat,
-  chatListRemoveChat,
-  // chatListRenameChat,
-  chatListSelectChat,
-} from "../chatList/state/chatListActions";
-import { chatSendMessageWithThunk } from "../chat/state/chatActions";
-import { selectChatMessages } from "../store/chatReducer/selectors";
-import { selectChatList } from "../store/chatListReducer/selectors";
-import { selectProfile } from "../store/profileReducer/selectors";
 
 const useStyles = makeStyles((theme) => ({
   rootGrid: {
@@ -27,82 +14,16 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-const getChatUrlById = (id) => {
-  return id != null ? `/chats/${id}` : "/chats";
-};
-
-function ChatsView() {
+function ChatsView({
+  chatList,
+  currentChatId,
+  handleChatSelect,
+  onAddNewChat,
+  onDelCurrentChat,
+  messageList,
+  onSendUserMessage,
+}) {
   const classes = useStyles();
-  const { chatId: urlChatId } = useParams();
-  const history = useHistory();
-  const goToChatUrlById = useCallback(
-    (id) => {
-      history.replace(getChatUrlById(id));
-    },
-    [history]
-  );
-  const dispatch = useDispatch();
-  const { chats: chatList, currentChatId } = useSelector(selectChatList);
-  const messageList = useSelector(selectChatMessages);
-  const { name } = useSelector(selectProfile);
-
-  const urlChatIdProvided = useMemo(
-    () => typeof urlChatId !== "undefined" && String(urlChatId) !== "",
-    [urlChatId]
-  );
-  const safeUrlChatId = useMemo(
-    () => (urlChatIdProvided ? String(urlChatId) : null),
-    [urlChatIdProvided, urlChatId]
-  );
-
-  // initial chat selection
-  useEffect(() => {
-    // console.log({ urlChatIdProvided, safeUrlChatId, currentChatId });
-    if (urlChatIdProvided) {
-      const referredChatExist = chatList.some(
-        (chat) => chat.id === safeUrlChatId
-      );
-      if (referredChatExist) {
-        if (safeUrlChatId !== currentChatId) {
-          // console.log("111");
-          dispatch(chatListSelectChat(safeUrlChatId));
-        }
-      } else {
-        // console.log("222");
-        history.replace(getChatUrlById(currentChatId));
-      }
-    } else if (currentChatId != null) {
-      // console.log("333");
-      history.replace(getChatUrlById(currentChatId));
-    }
-  });
-
-  const onSendUserMessage = useCallback(
-    (text) => {
-      const date = new Date();
-      const newMsg = {
-        author: String(name),
-        text,
-        date: date.toLocaleDateString(),
-        time: date.toLocaleTimeString(),
-      };
-      dispatch(chatSendMessageWithThunk(currentChatId, newMsg));
-    },
-    [dispatch, currentChatId, name]
-  );
-
-  const handleChatSelect = (id) => {
-    dispatch(chatListSelectChat(id));
-    goToChatUrlById(id);
-  };
-
-  const onAddNewChat = () => {
-    dispatch(chatListAddChat(`Chat ${chatList.length + 1}`));
-  };
-
-  const onDelCurrentChat = () => {
-    dispatch(chatListRemoveChat(currentChatId));
-  };
 
   return (
     <main className="ChatsView-main">
